requirements.txt
line-bot-sdk
flask
APScheduler
fuzzywuzzy
python-Levenshtein
gunicorn
gevent
requests
pyquery
python-dotenv
revChatGPT